wip: add pagination to sampling sites tables
