+++ app/src/contexts/surveyContext.tsx
@@ -1,12 +1,10 @@
 import { useBiohubApi } from 'hooks/useBioHubApi';
 import useDataLoader, { DataLoader } from 'hooks/useDataLoader';
 import { ICritterSimpleResponse } from 'interfaces/useCritterApi.interface';
-import { IGetSampleSiteNonSpatialResponse } from 'interfaces/useSamplingSiteApi.interface';
 import { IGetSurveyAttachmentsResponse, IGetSurveyForViewResponse } from 'interfaces/useSurveyApi.interface';
 import { IGetTechniquesResponse } from 'interfaces/useTechniqueApi.interface';
 import { createContext, PropsWithChildren, useEffect, useMemo } from 'react';
 import { useParams } from 'react-router';
-import { ApiPaginationRequestOptions } from 'types/misc';
 
 /**
  * Context object that stores information about a survey
@@ -30,18 +28,6 @@ export interface ISurveyContext {
    * @memberof ISurveyContext
    */
   artifactDataLoader: DataLoader<[project_id: number, survey_id: number], IGetSurveyAttachmentsResponse, unknown>;
-
-  /**
-   * The Data Loader used to load survey sample site data
-   *
-   * @type {DataLoader<[project_id: number, survey_id: number], IGetSampleSiteResponse, unknown>}
-   * @memberof ISurveyContext
-   */
-  sampleSiteDataLoader: DataLoader<
-    [project_id: number, survey_id: number, pagination?: ApiPaginationRequestOptions],
-    IGetSampleSiteNonSpatialResponse,
-    unknown
-  >;
 
   /**
    * The Data Loader used to load survey techniques
@@ -79,11 +65,6 @@ export interface ISurveyContext {
 export const SurveyContext = createContext<ISurveyContext>({
   surveyDataLoader: {} as DataLoader<[project_id: number, survey_id: number], IGetSurveyForViewResponse, unknown>,
   artifactDataLoader: {} as DataLoader<[project_id: number, survey_id: number], IGetSurveyAttachmentsResponse, unknown>,
-  sampleSiteDataLoader: {} as DataLoader<
-    [project_id: number, survey_id: number, pagination?: ApiPaginationRequestOptions],
-    IGetSampleSiteNonSpatialResponse,
-    unknown
-  >,
   techniqueDataLoader: {} as DataLoader<[project_id: number, survey_id: number], IGetTechniquesResponse, unknown>,
   critterDataLoader: {} as DataLoader<[project_id: number, survey_id: number], ICritterSimpleResponse[], unknown>,
   projectId: -1,
@@ -94,7 +75,6 @@ export const SurveyContextProvider = (props: PropsWithChildren<Record<never, any
   const biohubApi = useBiohubApi();
   const surveyDataLoader = useDataLoader(biohubApi.survey.getSurveyForView);
   const artifactDataLoader = useDataLoader(biohubApi.survey.getSurveyAttachments);
-  const sampleSiteDataLoader = useDataLoader(biohubApi.samplingSite.getSampleSites);
   const critterDataLoader = useDataLoader(biohubApi.survey.getSurveyCritters);
   const techniqueDataLoader = useDataLoader(biohubApi.technique.getTechniquesForSurvey);
 
@@ -117,7 +97,6 @@ export const SurveyContextProvider = (props: PropsWithChildren<Record<never, any
 
   surveyDataLoader.load(projectId, surveyId);
   artifactDataLoader.load(projectId, surveyId);
-  sampleSiteDataLoader.load(projectId, surveyId);
   critterDataLoader.load(projectId, surveyId);
 
   /**
@@ -132,7 +111,6 @@ export const SurveyContextProvider = (props: PropsWithChildren<Record<never, any
     ) {
       surveyDataLoader.refresh(projectId, surveyId);
       artifactDataLoader.refresh(projectId, surveyId);
-      sampleSiteDataLoader.refresh(projectId, surveyId);
       critterDataLoader.refresh(projectId, surveyId);
     }
 
@@ -143,21 +121,12 @@ export const SurveyContextProvider = (props: PropsWithChildren<Record<never, any
     return {
       surveyDataLoader,
       artifactDataLoader,
-      sampleSiteDataLoader,
       critterDataLoader,
       techniqueDataLoader,
       projectId,
       surveyId
     };
-  }, [
-    surveyDataLoader,
-    artifactDataLoader,
-    sampleSiteDataLoader,
-    critterDataLoader,
-    techniqueDataLoader,
-    projectId,
-    surveyId
-  ]);
+  }, [surveyDataLoader, artifactDataLoader, critterDataLoader, techniqueDataLoader, projectId, surveyId]);
 
   return <SurveyContext.Provider value={surveyContext}>{props.children}</SurveyContext.Provider>;
 };
