+++ app/src/features/surveys/sampling-information/sites/table/view/SamplingSiteTableView.tsx
@@ -18,27 +18,24 @@ interface ISamplingSiteManageTableView {
 
 export type ISamplingSiteCount = Record<SamplingSiteManageTableView, number>;
 
-interface ISamplingSiteTabsProps {
+interface ISamplingSiteTableViewProps {
   activeView: SamplingSiteManageTableView;
   setActiveView: React.Dispatch<SetStateAction<SamplingSiteManageTableView>>;
-  viewCounts: Record<SamplingSiteManageTableView, number>;
 }
 
 /**
  * Renders tab controls for the sampling site table, which allow the user to switch between viewing sites and periods.
  *
- * @param {ISamplingSiteTabsProps} props
+ * @param {ISamplingSiteTableViewProps} props
  * @return {*}
  */
-export const SamplingSiteTabs = (props: ISamplingSiteTabsProps) => {
-  const { activeView, setActiveView, viewCounts } = props;
+export const SamplingSiteTableView = (props: ISamplingSiteTableViewProps) => {
+  const { activeView, setActiveView } = props;
 
   const views: ISamplingSiteManageTableView[] = [
     { value: SamplingSiteManageTableView.SITES, icon: <Icon path={mdiMapMarker} size={0.75} /> },
     { value: SamplingSiteManageTableView.PERIODS, icon: <Icon path={mdiCalendarRange} size={0.75} /> }
   ];
-
-  const activeViewCount = viewCounts[activeView];
 
   const updateDatasetView = (_: React.MouseEvent<HTMLElement>, view: SamplingSiteManageTableView) => {
     if (view) {
@@ -67,7 +64,7 @@ export const SamplingSiteTabs = (props: ISamplingSiteTabsProps) => {
         }}>
         {views.map((view) => (
           <ToggleButton key={view.value} component={Button} color="primary" value={view.value} startIcon={view.icon}>
-            {view.value}&nbsp;({activeViewCount})
+            {view.value}
           </ToggleButton>
         ))}
       </ToggleButtonGroup>
