+++ app/src/features/surveys/view/components/sampling-data/components/SurveySamplingTabs.tsx
@@ -5,6 +5,7 @@ import Button from '@mui/material/Button';
 import Divider from '@mui/material/Divider';
 import ToggleButton from '@mui/material/ToggleButton';
 import ToggleButtonGroup from '@mui/material/ToggleButtonGroup';
+import { GridPaginationModel, GridSortModel } from '@mui/x-data-grid';
 import { LoadingGuard } from 'components/loading/LoadingGuard';
 import { SkeletonTable } from 'components/loading/SkeletonLoaders';
 import { NoDataOverlay } from 'components/overlay/NoDataOverlay';
@@ -13,15 +14,18 @@ import {
   SamplingPeriodTable
 } from 'features/surveys/sampling-information/periods/table/SamplingPeriodTable';
 import {
-  ISurveySitesRowData,
-  SurveySitesTable
-} from 'features/surveys/view/components/sampling-data/components/SurveySitesTable';
-import {
   ISurveyTechniqueRowData,
   SurveyTechniquesTable
 } from 'features/surveys/view/components/sampling-data/components/SurveyTechniquesTable';
+import { useBiohubApi } from 'hooks/useBioHubApi';
 import { useSurveyContext } from 'hooks/useContext';
+import useDataLoader from 'hooks/useDataLoader';
 import { useEffect, useMemo, useState } from 'react';
+import { ApiPaginationRequestOptions } from 'types/misc';
+import { firstOrNull } from 'utils/Utils';
+import { SurveySitesTable } from './SurveySitesTable';
+
+const pageSizeOptions = [10, 25, 50];
 
 export enum SurveySamplingView {
   TECHNIQUES = 'TECHNIQUES',
@@ -33,33 +37,34 @@ export const SurveySamplingTabs = () => {
   const surveyContext = useSurveyContext();
 
   const [activeView, setActiveView] = useState<SurveySamplingView>(SurveySamplingView.TECHNIQUES);
+  const biohubApi = useBiohubApi();
 
-  useEffect(() => {
-    // Refresh the data for the active view if the project or survey ID changes
-    if (activeView === SurveySamplingView.TECHNIQUES) {
-      surveyContext.techniqueDataLoader.refresh(surveyContext.projectId, surveyContext.surveyId);
-    }
-    if (activeView === SurveySamplingView.SITES) {
-      surveyContext.sampleSiteDataLoader.refresh(surveyContext.projectId, surveyContext.surveyId);
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [activeView]);
+  // useEffect(() => {
+  //   // Refresh the data for the active view if the project or survey ID changes
+  //   if (activeView === SurveySamplingView.TECHNIQUES) {
+  //     surveyContext.techniqueDataLoader.refresh(surveyContext.projectId, surveyContext.surveyId);
+  //   }
+  //   if (activeView === SurveySamplingView.SITES) {
+  //     surveyContext.sampleSiteDataLoader.refresh(surveyContext.projectId, surveyContext.surveyId);
+  //   }
+  //   // eslint-disable-next-line react-hooks/exhaustive-deps
+  // }, [activeView]);
 
-  useEffect(() => {
-    // Load the data initially once per tab, if/when the active view changes
-    if (activeView === SurveySamplingView.TECHNIQUES) {
-      surveyContext.techniqueDataLoader.load(surveyContext.projectId, surveyContext.surveyId);
-    }
-    if (activeView === SurveySamplingView.SITES) {
-      surveyContext.sampleSiteDataLoader.load(surveyContext.projectId, surveyContext.surveyId);
-    }
-  }, [
-    activeView,
-    surveyContext.techniqueDataLoader,
-    surveyContext.sampleSiteDataLoader,
-    surveyContext.projectId,
-    surveyContext.surveyId
-  ]);
+  // useEffect(() => {
+  //   // Load the data initially once per tab, if/when the active view changes
+  //   if (activeView === SurveySamplingView.TECHNIQUES) {
+  //     surveyContext.techniqueDataLoader.load(surveyContext.projectId, surveyContext.surveyId);
+  //   }
+  //   if (activeView === SurveySamplingView.SITES) {
+  //     surveyContext.sampleSiteDataLoader.load(surveyContext.projectId, surveyContext.surveyId);
+  //   }
+  // }, [
+  //   activeView,
+  //   surveyContext.techniqueDataLoader,
+  //   surveyContext.sampleSiteDataLoader,
+  //   surveyContext.projectId,
+  //   surveyContext.surveyId
+  // ]);
 
   const techniques: ISurveyTechniqueRowData[] =
     surveyContext.techniqueDataLoader.data?.techniques.map((technique) => ({
@@ -71,23 +76,42 @@ export const SurveySamplingTabs = () => {
       distance_threshold: technique.distance_threshold
     })) ?? [];
 
-  const sampleSites: ISurveySitesRowData[] = useMemo(
-    () =>
-      surveyContext.sampleSiteDataLoader.data?.sampleSites.map((site) => ({
-        id: site.survey_sample_site_id,
-        name: site.name,
-        description: site.description,
-        geometry_type: site.geometry_type,
-        blocks: site.blocks.map((block) => block.name),
-        stratums: site.stratums.map((stratum) => stratum.name)
-      })) ?? [],
-    [surveyContext.sampleSiteDataLoader.data?.sampleSites]
+  const [paginationModel, setPaginationModel] = useState<GridPaginationModel>({
+    page: 0,
+    pageSize: pageSizeOptions[0]
+  });
+  const [sortModel, setSortModel] = useState<GridSortModel>([]);
+
+  const samplingSitesDataLoader = useDataLoader((pagination: ApiPaginationRequestOptions) =>
+    biohubApi.samplingSite.getSampleSites(surveyContext.projectId, surveyContext.surveyId, pagination)
   );
 
+  const pagination: ApiPaginationRequestOptions = useMemo(() => {
+    const sort = firstOrNull(sortModel);
+
+    return {
+      limit: paginationModel.pageSize,
+      sort: sort?.field || undefined,
+      order: sort?.sort || undefined,
+
+      // API pagination pages begin at 1, but MUI DataGrid pagination begins at 0.
+      page: paginationModel.page + 1
+    };
+  }, [sortModel, paginationModel]);
+
+  // Refresh survey list when pagination or sort changes
+  useEffect(() => {
+    samplingSitesDataLoader.refresh(pagination);
+
+    // Adding a DataLoader as a dependency causes an infinite rerender loop if a useEffect calls `.refresh`
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [pagination]);
+
+  const sampleSites = samplingSitesDataLoader.data?.sampleSites ?? [];
   const samplePeriods: ISamplingSitePeriodRowData[] = useMemo(() => {
     const data: ISamplingSitePeriodRowData[] = [];
 
-    for (const site of surveyContext.sampleSiteDataLoader.data?.sampleSites ?? []) {
+    for (const site of sampleSites) {
       for (const method of site.sample_methods) {
         for (const period of method.sample_periods) {
           data.push({
@@ -105,11 +129,7 @@ export const SurveySamplingTabs = () => {
     }
 
     return data;
-  }, [surveyContext.sampleSiteDataLoader.data?.sampleSites]);
-
-  const techniquesCount = surveyContext.techniqueDataLoader.data?.count;
-  const sampleSitesCount = surveyContext.sampleSiteDataLoader.data?.sampleSites.length;
-  const samplePeriodsCount = samplePeriods.length;
+  }, [sampleSites]);
 
   return (
     <>
@@ -144,7 +164,7 @@ export const SurveySamplingTabs = () => {
             color="primary"
             startIcon={<Icon path={mdiAutoFix} size={0.75} />}
             value={SurveySamplingView.TECHNIQUES}>
-            {`${SurveySamplingView.TECHNIQUES} (${techniquesCount ?? 0})`}
+            {SurveySamplingView.TECHNIQUES}
           </ToggleButton>
           <ToggleButton
             key="sampling-sites-view"
@@ -152,7 +172,7 @@ export const SurveySamplingTabs = () => {
             color="primary"
             startIcon={<Icon path={mdiMapMarker} size={0.75} />}
             value={SurveySamplingView.SITES}>
-            {`${SurveySamplingView.SITES} (${sampleSitesCount ?? 0})`}
+            {SurveySamplingView.SITES}
           </ToggleButton>
           <ToggleButton
             key="sampling-sites-view"
@@ -160,7 +180,7 @@ export const SurveySamplingTabs = () => {
             color="primary"
             startIcon={<Icon path={mdiCalendarRange} size={0.75} />}
             value={SurveySamplingView.PERIODS}>
-            {`${SurveySamplingView.PERIODS} (${samplePeriodsCount ?? 0})`}
+            {SurveySamplingView.PERIODS}
           </ToggleButton>
         </ToggleButtonGroup>
       </Box>
@@ -169,67 +189,70 @@ export const SurveySamplingTabs = () => {
 
       <Box px={2} position="relative">
         {activeView === SurveySamplingView.TECHNIQUES && (
-          <>
-            <LoadingGuard
-              isLoading={surveyContext.techniqueDataLoader.isLoading || !surveyContext.techniqueDataLoader.isReady}
-              isLoadingFallback={<SkeletonTable />}
-              isLoadingFallbackDelay={100}
-              hasNoData={!techniquesCount}
-              hasNoDataFallback={
-                <NoDataOverlay
-                  height="250px"
-                  title="Add Techniques"
-                  subtitle="Techniques describe how you collected species observations"
-                  icon={mdiArrowTopRight}
-                />
-              }
-              hasNoDataFallbackDelay={100}>
-              <SurveyTechniquesTable techniques={techniques} />
-            </LoadingGuard>
-          </>
+          <LoadingGuard
+            isLoading={surveyContext.techniqueDataLoader.isLoading || !surveyContext.techniqueDataLoader.isReady}
+            isLoadingFallback={<SkeletonTable />}
+            isLoadingFallbackDelay={100}
+            hasNoData={!techniques.length}
+            hasNoDataFallback={
+              <NoDataOverlay
+                height="250px"
+                title="Add Techniques"
+                subtitle="Techniques describe how you collected species observations"
+                icon={mdiArrowTopRight}
+              />
+            }
+            hasNoDataFallbackDelay={100}>
+            <SurveyTechniquesTable techniques={techniques} />
+          </LoadingGuard>
         )}
 
+        {/* Data tables */}
+
         {activeView === SurveySamplingView.SITES && (
-          <>
-            <LoadingGuard
-              isLoading={surveyContext.sampleSiteDataLoader.isLoading || !surveyContext.sampleSiteDataLoader.isReady}
-              isLoadingFallback={<SkeletonTable />}
-              isLoadingFallbackDelay={100}
-              hasNoData={!sampleSitesCount}
-              hasNoDataFallback={
-                <NoDataOverlay
-                  height="250px"
-                  title="Add Sampling Sites"
-                  subtitle="Apply your techniques to sampling sites to show where you collected data"
-                  icon={mdiArrowTopRight}
-                />
-              }
-              hasNoDataFallbackDelay={100}>
-              <SurveySitesTable sites={sampleSites} />
-            </LoadingGuard>
-          </>
+          <LoadingGuard
+            isLoading={samplingSitesDataLoader.isLoading || !samplingSitesDataLoader.isReady}
+            isLoadingFallback={<SkeletonTable />}
+            isLoadingFallbackDelay={100}
+            hasNoData={!sampleSites.length}
+            hasNoDataFallback={
+              <NoDataOverlay
+                height="200px"
+                title="Add Sampling Sites"
+                subtitle="Apply your techniques to sampling sites to show where you collected data"
+                icon={mdiArrowTopRight}
+              />
+            }
+            hasNoDataFallbackDelay={100}>
+            <SurveySitesTable
+              sites={sampleSites}
+              setPaginationModel={setPaginationModel}
+              setSortModel={setSortModel}
+              paginationModel={paginationModel}
+              pageSizeOptions={pageSizeOptions}
+              sortModel={sortModel}
+              handleRefresh={() => {}}
+            />
+          </LoadingGuard>
         )}
 
         {activeView === SurveySamplingView.PERIODS && (
-          <>
-            <LoadingGuard
-              isLoading={surveyContext.sampleSiteDataLoader.isLoading || !surveyContext.sampleSiteDataLoader.isReady}
-              isLoadingFallback={<SkeletonTable />}
-              isLoadingFallbackDelay={100}
-              hasNoData={!samplePeriodsCount}
-              hasNoDataFallback={
-                <NoDataOverlay
-                  height="250px"
-                  title="Add Periods"
-                  subtitle="Add periods when you create sampling sites to show when 
-                  you collected species observations"
-                  icon={mdiArrowTopRight}
-                />
-              }
-              hasNoDataFallbackDelay={100}>
-              <SamplingPeriodTable periods={samplePeriods} />
-            </LoadingGuard>
-          </>
+          <LoadingGuard
+            isLoading={samplingSitesDataLoader.isLoading || !samplingSitesDataLoader.isReady}
+            isLoadingFallback={<SkeletonTable />}
+            isLoadingFallbackDelay={100}
+            hasNoData={!samplePeriods.length}
+            hasNoDataFallback={
+              <NoDataOverlay
+                height="200px"
+                title="Add Periods"
+                subtitle="Add periods when you create sampling sites to show when you collected species observations"
+                icon={mdiArrowTopRight}
+              />
+            }
+            hasNoDataFallbackDelay={100}>
+            <SamplingPeriodTable periods={samplePeriods} />
+          </LoadingGuard>
         )}
       </Box>
     </>
